# agent_api_http/openapi-full.yaml
openapi: 3.1.0
info:
  title: UniCore HTTP API
  description: ''
  license:
    name: Apache 2.0
  version: 0.0.0-semantically-released
servers:
- url: http://localhost:3033
  description: Local development
paths:
...
  /openid4vci/nonce:
    post:
      tags:
      - OpenID4VCI
      - Protocol
      summary: Get a fresh nonce
      description: |-
        Returns a fresh `c_nonce` to be used in key proofs, as defined by
        [OpenID4VCI](https://openid.net/specs/openid-4-verifiable-credential-issuance-1_0.html#name-nonce-endpoint).
      operationId: openid4vci_nonce
      responses:
        '200':
          description: Fresh nonce
          headers:
            Cache-Control:
              schema:
                type: string
              description: Always `no-store`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NonceResponse'
        '500':
          description: The nonce could not be generated
...
  /redirect:
    post:
      tags:
      - OID4VP / SIOPv2
      - Protocol
      summary: Submit an authorization response
      description: |-
        Receives and verifies the authorization response of a SIOPv2 or OID4VP authorization request, sent with the
        `direct_post` response mode.
      operationId: redirect
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AuthorizationResponseForm'
        required: true
      responses:
        '200':
          description: Authorization response verified
        '400':
          description: The authorization response is malformed or has no `state`
        '500':
          description: The authorization response could not be verified
...
  /vct/{credential_configuration_id}/{version}:
    get:
      tags:
      - SD-JWT VC
      - Protocol
      summary: Get SD-JWT VC type metadata
      description: |-
        Returns the type metadata of a credential configuration, as defined by
        [SD-JWT VC](https://datatracker.ietf.org/doc/html/draft-ietf-oauth-sd-jwt-vc#name-sd-jwt-vc-type-metadata).
      operationId: vct_type_metadata
      parameters:
      - name: credential_configuration_id
        in: path
        description: Base64url-encoded (without padding) credential configuration ID
        required: true
        schema:
          type: string
      - name: version
        in: path
        description: Type version (currently ignored)
        required: true
        schema:
          type: string
      responses:
        '200':
          description: SD-JWT VC type metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TypeMetadata'
        '404':
          description: The credential configuration does not exist
        '500':
          description: The type metadata could not be retrieved
...
components:
  schemas:
...
    AuthorizationResponseForm:
      oneOf:
      - $ref: '#/components/schemas/IdTokenAuthorizationResponse'
      - $ref: '#/components/schemas/VpTokenAuthorizationResponse'
      description: An authorization response sent to the relying party, either as a SIOPv2 ID Token or as an OID4VP VP Token.
...
    ClaimPathPointer:
      type: array
      items:
        oneOf:
        - type: string
          description: To address a particular claim within an object, append the key (claim name) to the array.
        - type: integer
          format: int64
          description: To address an element within an array, append the index to the array (as a non-negative, 0-based integer).
          minimum: 0
        - type: 'null'
          description: To address all elements within an array, append a null value to the array.
          default: null
        description: One segment of a Claim Path Pointer.
      minItems: 1
...
    IdTokenAuthorizationResponse:
      type: object
      description: |-
        A [SIOPv2](https://openid.net/specs/openid-connect-self-issued-v2-1_0.html#name-self-issued-openid-provider-a)
        authorization response.
      required:
      - state
      - id_token
      properties:
        id_token:
          type: string
          description: The Self-Issued ID Token as a compact JWT.
        state:
          type: string
          description: The `state` of the authorization request, which identifies it.
...
    NonceResponse:
      type: object
      description: |-
        Nonce Response as defined by OpenID4VCI.

        See <https://openid.net/specs/openid-4-verifiable-credential-issuance-1_0.html#name-nonce-response>.
      required:
      - c_nonce
      properties:
        c_nonce:
          type: string
...
    TypeMetadata:
      type: object
      description: |-
        OpenAPI representation of `identity_credential::sd_jwt_vc::metadata::TypeMetadata`.

        See [SD-JWT VC Type Metadata](https://datatracker.ietf.org/doc/html/draft-ietf-oauth-sd-jwt-vc#name-sd-jwt-vc-type-metadata).
      properties:
        claims:
          type: array
          items:
            $ref: '#/components/schemas/TypeMetadataClaim'
        description:
          type:
          - string
          - 'null'
          description: A human-readable description for the type.
        display:
          type: array
          items:
            $ref: '#/components/schemas/TypeMetadataDisplay'
        extends:
          type:
          - string
          - 'null'
          format: uri
          description: A URI of another type that this type extends.
        extends#integrity:
          type:
          - string
          - 'null'
          description: Integrity metadata for the extended type.
        name:
          type:
          - string
          - 'null'
          description: A human-readable name for the type.
        schema:
          type:
          - object
          - 'null'
          description: An embedded JSON Schema for the credential.
          additionalProperties: {}
          propertyNames:
            type: string
        schema_uri:
          type:
          - string
          - 'null'
          format: uri
          description: A URI referencing a JSON Schema for the credential.
        schema_uri#integrity:
          type:
          - string
          - 'null'
          description: Integrity metadata for the referenced JSON Schema.
    TypeMetadataClaim:
      type: object
      description: Information about particular claims of an SD-JWT VC type.
      required:
      - path
      properties:
        display:
          type: array
          items:
            $ref: '#/components/schemas/TypeMetadataClaimDisplay'
        mandatory:
          type:
          - boolean
          - 'null'
          description: Whether the claim must be present in the issued credential.
        path:
          $ref: '#/components/schemas/ClaimPathPointer'
        sd:
          oneOf:
          - type: 'null'
          - type: string
            enum:
            - always
            - allowed
            - never
          description: Whether the claim is selectively disclosable.
        svg_id:
          type:
          - string
          - 'null'
          description: The ID of the claim for reference in an SVG template.
    TypeMetadataClaimDisplay:
      type: object
      description: Display information for a claim.
      required:
      - locale
      - label
      properties:
        description:
          type:
          - string
          - 'null'
        label:
          type: string
        locale:
          type: string
          description: A language tag as defined in [RFC 5646](https://www.rfc-editor.org/rfc/rfc5646.txt).
    TypeMetadataDisplay:
      type: object
      description: Display information for an SD-JWT VC type.
      required:
      - locale
      - name
      properties:
        description:
          type:
          - string
          - 'null'
        locale:
          type: string
          description: A language tag as defined in [RFC 5646](https://www.rfc-editor.org/rfc/rfc5646.txt).
        name:
          type: string
        rendering:
          type:
          - object
          - 'null'
          description: Rendering information, such as `simple` or `svg_templates`.
          additionalProperties: {}
          propertyNames:
            type: string
...
    VpTokenAuthorizationResponse:
      type: object
      description: |-
        An [OID4VP](https://openid.net/specs/openid-4-verifiable-presentations-1_0.html#name-response) authorization
        response.
      required:
      - state
      - vp_token
      properties:
        state:
          type: string
          description: The `state` of the authorization request, which identifies it.
        vp_token:
          type: string
          description: JSON-encoded object that maps each DCQL credential query ID to a non-empty array of presentations.
  securitySchemes:
    access_token:
      type: http
      scheme: bearer
      description: Access token issued by the `/auth/token` endpoint.
...